requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
flask==3.1.0
flask-mail==0.10.0
gunicorn==23.0.0
itsdangerous==2.2.0
jinja2==3.1.6
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
mysql-connector-python==9.3.0
packaging==25.0
pycparser==2.22
pymysql==1.1.1
python-dotenv==1.1.0
werkzeug==3.1.3
